app/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.4.0
flake8==6.0.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
openapi-codec==1.3.2
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.3.1
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
